+++ src/pages/Home.jsx
@@ -1,13 +1,21 @@
 import React, { useEffect } from 'react';
+import { useLocation } from 'react-router-dom';
 import Navbar from '../components/Navbar/Navbar';
 import Hero from '../components/Hero/Hero';
-// import About from '../components/About/About';
-import Skillset from '../components/Skillset/Skillset';
-import Socials from '../components/Socials/Socials';
+import HomeAbout from '../components/HomeAbout/HomeAbout';
 import Projects from '../components/Projects/Projects';
-import Contact from '../components/Contact/Contact';
+import Workshop from '../components/Workshop/Workshop';
+import Gaming from '../components/Gaming/Gaming';
+import Connect from '../components/Connect/Connect';
 import Footer from '../components/Footer/Footer';
-import { useLocation } from 'react-router-dom';
+
+// import Hero from '../components/Hero/Hero';
+// import About from '../components/About/About';
+// import Skillset from '../components/Skillset/Skillset';
+// import Socials from '../components/Socials/Socials';
+// import Contact from '../components/Contact/Contact';
+// import Footer from '../components/Footer/Footer';
+
 
 const Home = () => {
   const { hash } = useLocation();
@@ -27,17 +35,31 @@ const Home = () => {
 
   return (
     <div>
-      <header>
+      <header className='main'>
+        <Navbar isHome={true} />
+      </header>
+      <main>
+        <Hero />
+        <HomeAbout />
+        <Workshop />
+        <Projects />
+        <Gaming />
+        <Connect />
+      </main>
+      <Footer />
+
+
+      {/* <header>
         <Navbar isHome={true} />
       </header>
       <main>
         <Hero />
         <Skillset />
-        <Projects />
+        
         <Socials />
         <Contact />
       </main>
-      <Footer />
+      <Footer /> */}
     </div>
   )
 }