+++ src/pages/Terms.jsx
@@ -1,7 +1,6 @@
 import React, { useEffect } from 'react'
 import Navbar from "../components/Navbar/Navbar";
 import Footer from "../components/Footer/Footer";
-import './Pages.scss';
 
 const Terms = () => {