+++ CHANGELOG.md
@@ -10,26 +10,33 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ---
 
-## [0.2.5] - 2024-09-02
+## [0.2.5] - 2024-09-05
 ### Fixed
 - [About] Social media display error
 - [Navbar] Added support for navigating between page-to-anchor link to another page via "isHome" prop
+- [Global] Fixed duplicated styles on SCSS compile
 ### Added
-- [Home] Added photo flipcard
 - [About] PH language support
 - [Twitch] Embedded player
 - [Twitch] History
 - [Twitch] Now Streaming + Community
 - [Projects] added
-- [Socials] added
+- [Workshop] added
+- [Gaming] added
+- [Connect] added
 ### Changed
 - [Global] Cleaned up global styles
+- [Global] Changed colour schemes
 - [Global] Fixed pages to top scroll for each load
 - [About] Updated/cleaned layout
 - [Home] Updated layout
+- [Hero] revamp
 - [Projects] renamed from [MyWork]
+- [Projects] revamp
 ### Removed
 - [Services] removed
+- [Socials] removed
+- cleaned up unused assets/components
 
 ## [0.2.4] - 2024-08-28
 ### Added